FEAT #9156 Admin list display

+++ src/app/basket/models/GroupBasketModelAbstract.php
@@ -55,19 +55,36 @@ abstract class GroupBasketModelAbstract
         return true;
     }
 
+    public static function update(array $aArgs)
+    {
+        ValidatorModel::notEmpty($aArgs, ['set', 'where', 'data']);
+        ValidatorModel::arrayType($aArgs, ['set', 'where', 'data']);
+
+        DatabaseModel::update([
+            'table' => 'groupbasket',
+            'set'   => $aArgs['set'],
+            'where' => $aArgs['where'],
+            'data'  => $aArgs['data']
+        ]);
+
+        return true;
+    }
+
     public static function deleteGroupBasket(array $aArgs)
     {
         ValidatorModel::notEmpty($aArgs, ['basketId', 'groupId']);
         ValidatorModel::stringType($aArgs, ['basketId', 'groupId']);
-        ValidatorModel::boolType($aArgs, ['preferences']);
+        ValidatorModel::boolType($aArgs, ['preferences', 'groupBasket']);
 
         $group = GroupModel::getByGroupId(['select' => ['id'], 'groupId' => $aArgs['groupId']]);
 
-        DatabaseModel::delete([
-            'table' => 'groupbasket',
-            'where' => ['basket_id = ?', 'group_id = ?'],
-            'data'  => [$aArgs['basketId'], $aArgs['groupId']]
-        ]);
+        if (!empty($aArgs['groupBasket'])) {
+            DatabaseModel::delete([
+                'table' => 'groupbasket',
+                'where' => ['basket_id = ?', 'group_id = ?'],
+                'data'  => [$aArgs['basketId'], $aArgs['groupId']]
+            ]);
+        }
         DatabaseModel::delete([
             'table' => 'actions_groupbaskets',
             'where' => ['basket_id = ?', 'group_id = ?'],
